Add an Ansible inventory for the ALaSKA Alt-1 partition
This change adds two new Ansible groups - alaska and alaska-alt-1. In the
existing inventory file, ansible/inventory, all hosts are added to the
alaska group. A second inventory file is added, ansible/inventory-alt-1,
which adds all hosts to the alaska-alt-1 group.

This allows us to specify different configuration for each ALaSKA environment,
through the use of group variables.

Support has been added to the stackhpc.cluster-infra and
stackhpc.os-container-infra roles for specifying a top level 'environment'
group to which all hosts are added. This ensures that the generated Ansible
inventories can also reference environment-specific variables.

A new variable, alaska_environment, specifies the name of the environment
group - alaska or alaska-alt-1.

Variables that have been determined to need to vary between environments are:

* controller_vip
* alaska_monitoring_server
* os_manila_mount_share_user
* os_manila_mount_os_config_name
* os_container_infra_cloud