Add test scripts and configuration for openhpc, swarm-sip, k8s
These scripts can be used to deploy clusters configured with a minimal number
of nodes (2), such that they are suitable for testing.

The scripts use the OS_CLOUD environment variable to detect the target
environment, and use an appropriate Ansible inventory.

This allows us to easily verify at a basic level that the cloud is still
functioning as expected.

Further work could be done to verify all the moving parts, including access to
shared storage services, and proper functioning of the batch scheduler or
container cluster.